A quantum machine learning algorithm for classifying MRI images of brain tumors.
This code uses the quantum support vector machine algorithm to classify MRI images of brain tumors. The quantum support vector machine algorithm is a quantum algorithm that can find the optimal hyperplane that separates two classes of data using a quantum feature map and a kernel computation. This is relevant to MRI because it can help diagnose malignant or benign tumors based on their pixel values and labels. The quantum support vector machine algorithm works by encoding each pixel value into a quantum state using rotations and entanglements. Then it uses a SWAP test to measure the overlap between two quantum states and computes the kernel value between them. The algorithm then solves a quadratic programming problem using cvxopt to find the optimal values of the variables that define the hyperplane. The algorithm then finds the support vectors and their labels and coefficients from the optimal values of the variables. The algorithm then computes the bias term for the hyperplane using quantum kernel computation. The algorithm then classifies a new data point by computing its kernel value with all support vectors and predicting its label based on the sign of its weighted sum.

I made it with the help of several DNN AI models. It is programed in Qiskit and is intended to be ran on a quantum computer or simulator. 

You may use and modify this machine learning algorithm for research and education or entertainment purposes provided you cite me as a source and give credit to my sources as well. 

This is not intended to be used for medical or commercial purposes without obtaining a licensing agreement from me and it should be noted that the code requires further modification and a large training dataset before this algorithm could ever be used in any medical capacity.

Inspired by Quantum Computers in the Revolution of Artificial Intelligence and Machine Learning and Quantum support vector machine